% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmr.R
\name{asmr}
\alias{asmr}
\title{summarise and classify asmr}
\usage{
asmr(stnData, bgc, soils = soilsData, future = TRUE,
  years = "random", lat = 55)
}
\arguments{
\item{stnData}{climate station data}

\item{bgc}{corresponding BGC variant for stnData (e.g., "SBSmk1)}

\item{soils}{default or user-defined soils data}

\item{future}{TRUE/FALSE if TRUE(default), then future scenarios are run for bgc}

\item{years}{either "random" or custom vector of years, as integers}

\item{lat}{latitude in degrees, either 50,55 or 60}
}
\description{
This function will summarise monthly asmr values and classes for current and future periods.
}
\details{
The function uses output from asmrRun():

Input:
}
\examples{
asmr(PrinceGeorge,bgc="SBSmk1")

}
