% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmrCalc.R
\name{asmrCalc}
\alias{asmrCalc}
\title{ASMR calculation}
\usage{
asmrCalc(climateData, latitude = 55, soilsData)
}
\arguments{
\item{climateData}{Climate data must be imported correctly.}
}
\description{
This function uses daily climate and soils data to estimate daily
water balance for five different soil moisture regimes.
}
