% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmrClass.R
\docType{data}
\name{asmrClass}
\alias{asmrClass}
\title{ASMR category data for AET/PET ratios.}
\format{A data frame with 6 rows and 2 variables
\describe{
\item{asmrUL}{asmrUL: Upper limit of AET/PET ratio for that category}
\item{class}{class: ASMR class}
  ...
}}
\usage{
asmrClass
}
\description{
Use this dataset when classifying AET/PET ratios into ASMR categories.
}
\keyword{datasets}
