% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmrNorth.R
\docType{data}
\name{asmrNorth}
\alias{asmrNorth}
\title{ASMR data for BGC units in northern BC}
\format{An object of class \code{list} of length 3.}
\usage{
asmrNorth
}
\description{
A list of current and future asmr values and classes.
}
\details{
* ...
}
\keyword{datasets}
