% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmrRun.R
\name{asmrRun}
\alias{asmrRun}
\title{run current and future ASMR}
\usage{
asmrRun(stnData, soils = soilsData, future = TRUE, bgc,
  years = "random", lat = 55)
}
\arguments{
\item{stnData}{climate station data}

\item{soils}{default or user-defined soils data}

\item{future}{TRUE/FALSE if TRUE(default), then future scenarios are run for bgc}

\item{bgc}{corresponding BGC variant for stnData (e.g., "SBSmk1)}

\item{years}{either "random" or custom vector of years, as integers}

\item{lat}{latitude in degrees, either 50,55 or 60}
}
\description{
This function will generate daily asmr values for current and future periods.
}
\details{
The function uses the following approach:

Input:
}
\examples{
asmrRun(PrinceGeorge,bgc="SBSmk1")
}
\keyword{internal}
