% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgcClim.R
\docType{data}
\name{bgcClim}
\alias{bgcClim}
\title{Current and future climate data for BGC units in British Columbia}
\format{A data frame with 1470 rows and 39 variables
\describe{
\item{BGC: biogeoclimatic unit}
\item{period: climatic period, either normal or future}
\item{Scenario: carbon emission scenario for future periods}
\item{assorted climate variables}
}}
\usage{
bgcClim
}
\description{
A dataset containing normal and future climate data.
Data are from ClimateBC (5.6)
}
\keyword{datasets}
