% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanECData.R
\name{cleanECData}
\alias{cleanECData}
\title{Clean Environment Canada daily climate data}
\usage{
cleanECData(climateData)
}
\arguments{
\item{climateData}{Climate data must be imported correctly.  Need to have year, month, day, tmn, tmx and ppt columns.}
}
\description{
This function takes a pre-formatted Environment Canada daily climate dataset
and prepares it for use in the asmrCalc() function.
}
\examples{

require(weathercan)
smithersData<-
weathercan::weather_dl(station_ids=487,interval="day",start="1955-01-01",end="1995-12-31") \%>\%
select(stn=station_name,date,tmn=min_temp,tmx=max_temp,ppt=total_precip,year,month,day)
cleanECData(smithersData)
}
