% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmrFuture.R
\name{futureASMR}
\alias{futureASMR}
\title{Calculate future ASMR}
\usage{
futureASMR(stnData, bgc)
}
\arguments{
\item{stnData}{climate station data}

\item{bgc}{corresponding BGC variant for stnData (e.g., "SBSmk1)}
}
\description{
This function will adjust observed daily data for future periods.
}
\details{
The function uses the following approach:

We use the normal (1961-1990) monthly data from ClimateBC (see ?bgcClim) to establish a baseline climate
for a given BGC unit.  The difference between normal period data and future data is calculated, and then
and the difference, expressed in percentage for precipitation and degrees for temperatures, is then applied to the daily data for a given climate station.

Future scenarios are calculated for three scenarios:
\itemize{
  \item RCP 4.5
  \item RCP 8.5
  \item Mean of RCP 4.5 and 8.5
}

and three time periods:
\itemize{
  \item 2025
  \item 2055
  \item 2085
}

Input:
}
\examples{
futureASMR(PrinceGeorge,"SBSmk1")

}
\keyword{internal}
