% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrinceGeorge.R
\docType{data}
\name{PrinceGeorge}
\alias{PrinceGeorge}
\title{Prince George daily climate data.}
\format{A data frame with 33215 rows and 7 variables:
\describe{
\item{date}{daily date}
\item{tmn}{daily minimum temperature, in C}
\item{tmx}{daily maximum temperature, in C}
\item{ppt}{daily precipitation, in mm}
\item{year}{year}
\item{month}{month}
\item{day}{day}
...
}}
\source{
\url{ftp://ccrp.tor.ec.gc.ca/pub/EC_data/AHCCD_daily/}
}
\usage{
PrinceGeorge
}
\description{
A dataset of adjusted and homogenized daily climate data from the Prince George airport station.
}
\keyword{datasets}
