% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseASMR.R
\name{summariseASMR}
\alias{summariseASMR}
\title{Summarise monthly ASMR values}
\usage{
summariseASMR(asmrData)
}
\arguments{
\item{asmrData}{Output from asmrCalc() function.}
}
\description{
This function summarises monthly ASMR values from the
asmrCalc() function.
}
\examples{
summariseASMR((asmrCalc(PrinceGeorge))

}
