% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeRisk.R
\docType{data}
\name{treeRisk}
\alias{treeRisk}
\title{Tree risk ratings for species in BC.}
\format{A data frame with 9 rows and 5 variables
\describe{
\item{Lower limit}{Lower asmr limit where species is found (needs to be updated)}
\item{High}{Tree is at high risk of drought mortality below this asmr}
\item{Moderate}{Below this asmr, and above high, tree is at moderate risk of drought mortality}
\item{Low}{Upper asmr limit where species is found.  Below this limit, to moderate, tree is at low risk of drought mortality.}
* ...
}}
\usage{
treeRisk
}
\description{
A dataset containing soils data to be used when calculating tree drought risk.
}
\keyword{datasets}
