% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearSelect.R
\name{yearSelect}
\alias{yearSelect}
\title{Select years to use in ASMR summaries}
\usage{
yearSelect(stnData, excl = c(1940, 2000), win = c(1961, 1990),
  yrs = 10)
}
\arguments{
\item{stnData}{cleaned climate data, properly formatted}

\item{excl}{data boundaries (usually 1940-2000, doesn`t need to be set)}

\item{win}{ideal climate normal period (usually 1961-1990, doesn`t need to be set)}

\item{yrs}{number of years to randomly select (usually 10, doesn`t need to be set)}
}
\description{
This function implements a random approach to selecting 10 years to pass to
summariseASMR() function.
}
\details{
The function uses the following approach:

\itemize{
  \item Data from 1961-1990 climate normal period are preferred
  \item No data from outside 1940-2000 period are used
  \item 10 years are selected randomly to generate ASMR summaries.  If there are 10+ years within the 1961-1990 period, years will be randomly selected from that time period.

}
}
\examples{
asmrSelect(PrinceGeorge)

}
\keyword{internal}
