% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrinceGeorge.R
\docType{data}
\name{PrinceGeorge}
\alias{PrinceGeorge}
\title{Daily climate data for Prince George Environment Canada climate station.}
\format{A data frame with 33215 rows and 7 variables}
\usage{
PrinceGeorge
}
\description{
A dataset containing daily temperature and precipitation data from 1918 to 2008.
}
\details{
The variables are as follows:
* `date`: YYYY-MM-DD
* `tmn`: daily minimum temperature, in Degree Celsius
* `tmx`: daily maximum temperature, in Degree Celsius
* `ppt`: daily precipitation amount, in mm
* ...
}
\keyword{datasets}
